+++ sdk/azcore/internal/shared/constants.go
@@ -35,5 +35,5 @@ const (
 	Module = "azcore"
 
 	// Version is the semantic version (see http://semver.org) of this module.
-	Version = "v0.20.0"
+	Version = "v0.20.1"
 )