+++ sdk/azidentity/client_certificate_credential_test.go
@@ -40,10 +40,6 @@ func TestClientCertificateCredential_InvalidTenantID(t *testing.T) {
 	if cred != nil {
 		t.Fatalf("Expected a nil credential value. Received: %v", cred)
 	}
-	var errType *CredentialUnavailableError
-	if !errors.As(err, &errType) {
-		t.Fatalf("Did not receive a CredentialUnavailableError. Received: %t", err)
-	}
 }
 
 func TestClientCertificateCredential_CreateAuthRequestSuccess(t *testing.T) {
@@ -166,7 +162,7 @@ func TestClientCertificateCredential_GetTokenSuccess(t *testing.T) {
 			srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
 			options := ClientCertificateCredentialOptions{}
 			options.AuthorityHost = AuthorityHost(srv.URL())
-			options.HTTPClient = srv
+			options.Transport = srv
 			options.Password = test.password
 			cred, err := NewClientCertificateCredential(tenantID, clientID, test.certData, &options)
 			if err != nil {
@@ -189,7 +185,7 @@ func TestClientCertificateCredential_GetTokenSuccess_withCertificateChain(t *tes
 			options := ClientCertificateCredentialOptions{}
 			options.AuthorityHost = AuthorityHost(srv.URL())
 			options.SendCertificateChain = true
-			options.HTTPClient = srv
+			options.Transport = srv
 			options.Password = test.password
 			cred, err := NewClientCertificateCredential(tenantID, clientID, test.certData, &options)
 			if err != nil {
@@ -211,7 +207,7 @@ func TestClientCertificateCredential_GetTokenInvalidCredentials(t *testing.T) {
 			srv.SetResponse(mock.WithStatusCode(http.StatusUnauthorized))
 			options := ClientCertificateCredentialOptions{}
 			options.AuthorityHost = AuthorityHost(srv.URL())
-			options.HTTPClient = srv
+			options.Transport = srv
 			options.Password = test.password
 			cred, err := NewClientCertificateCredential(tenantID, clientID, test.certData, &options)
 			if err != nil {
@@ -221,7 +217,7 @@ func TestClientCertificateCredential_GetTokenInvalidCredentials(t *testing.T) {
 			if err == nil {
 				t.Fatalf("Expected to receive a nil error, but received: %v", err)
 			}
-			var authFailed *AuthenticationFailedError
+			var authFailed AuthenticationFailedError
 			if !errors.As(err, &authFailed) {
 				t.Fatalf("Expected: AuthenticationFailedError, Received: %T", err)
 			}
@@ -235,7 +231,7 @@ func TestClientCertificateCredential_GetTokenCheckPrivateKeyBlocks(t *testing.T)
 	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
 	options := ClientCertificateCredentialOptions{}
 	options.AuthorityHost = AuthorityHost(srv.URL())
-	options.HTTPClient = srv
+	options.Transport = srv
 	certData, err := os.ReadFile("testdata/certificate_formatB.pem")
 	if err != nil {
 		t.Fatalf("Failed to read certificate file: %s", err.Error())
@@ -256,7 +252,7 @@ func TestClientCertificateCredential_NoData(t *testing.T) {
 	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
 	options := ClientCertificateCredentialOptions{}
 	options.AuthorityHost = AuthorityHost(srv.URL())
-	options.HTTPClient = srv
+	options.Transport = srv
 	_, err := NewClientCertificateCredential(tenantID, clientID, []byte{}, &options)
 	if err == nil {
 		t.Fatalf("Expected an error but received nil")
@@ -269,7 +265,7 @@ func TestClientCertificateCredential_NoCertificate(t *testing.T) {
 	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
 	options := ClientCertificateCredentialOptions{}
 	options.AuthorityHost = AuthorityHost(srv.URL())
-	options.HTTPClient = srv
+	options.Transport = srv
 	certData, err := os.ReadFile("testdata/certificate_empty.pem")
 	if err != nil {
 		t.Fatalf("Failed to read certificate file: %s", err.Error())
@@ -286,7 +282,7 @@ func TestClientCertificateCredential_NoPrivateKey(t *testing.T) {
 	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
 	options := ClientCertificateCredentialOptions{}
 	options.AuthorityHost = AuthorityHost(srv.URL())
-	options.HTTPClient = srv
+	options.Transport = srv
 	certData, err := os.ReadFile("testdata/certificate_nokey.pem")
 	if err != nil {
 		t.Fatalf("Failed to read certificate file: %s", err.Error())
@@ -304,7 +300,7 @@ func TestBearerPolicy_ClientCertificateCredential(t *testing.T) {
 	srv.AppendResponse(mock.WithStatusCode(http.StatusOK))
 	options := ClientCertificateCredentialOptions{}
 	options.AuthorityHost = AuthorityHost(srv.URL())
-	options.HTTPClient = srv
+	options.Transport = srv
 	cred, err := NewClientCertificateCredential(tenantID, clientID, pemCert, &options)
 	if err != nil {
 		t.Fatalf("Did not expect an error but received: %v", err)