+++ sdk/data/azcosmos/cosmos_headers.go
@@ -4,7 +4,6 @@
 package azcosmos
 
 const (
-	cosmosHeaderRetryAfter             string = "x-ms-retryafter"
 	cosmosHeaderRequestCharge          string = "x-ms-request-charge"
 	cosmosHeaderActivityId             string = "x-ms-activity-id"
 	cosmosHeaderEtag                   string = "etag"