+++ sdk/data/azcosmos/throughput_properties_test.go
@@ -7,6 +7,8 @@ import (
 	"encoding/json"
 	"testing"
 	"time"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 )
 
 func TestThroughputPropertiesManualRawSerialization(t *testing.T) {
@@ -36,17 +38,17 @@ func TestThroughputPropertiesManualRawSerialization(t *testing.T) {
 		t.Errorf("OfferId mismatch %v", otherProperties.offerId)
 	}
 
-	if otherProperties.ETag != "\"00000000-0000-0000-9b8c-8ea3e19601d7\"" {
+	if *otherProperties.ETag != "\"00000000-0000-0000-9b8c-8ea3e19601d7\"" {
 		t.Errorf("Etag mismatch %v", otherProperties.ETag)
 	}
 
-	if otherProperties.LastModified.Time != nowAsUnix {
-		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified.Time)
+	if otherProperties.LastModified != nowAsUnix {
+		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified)
 	}
 
-	mt, err := otherProperties.ManualThroughput()
-	if err != nil {
-		t.Fatal(err)
+	mt, isManual := otherProperties.ManualThroughput()
+	if !isManual {
+		t.Fatal("Expected to have manual throughput available")
 	}
 
 	if mt != 400 {
@@ -55,18 +57,15 @@ func TestThroughputPropertiesManualRawSerialization(t *testing.T) {
 }
 
 func TestThroughputPropertiesManualE2ESerialization(t *testing.T) {
-	nowAsUnix := time.Now().Unix()
+	nowAsUnix := time.Unix(time.Now().Unix(), 0)
 
-	now := UnixTime{
-		Time: time.Unix(nowAsUnix, 0),
-	}
-
+	etag := azcore.ETag("\"00000000-0000-0000-9b8c-8ea3e19601d7\"")
 	properties := NewManualThroughputProperties(400)
 	properties.offerId = "HFln"
 	properties.offerResourceId = "4SRTANCD3Dw="
-	properties.ETag = "\"00000000-0000-0000-9b8c-8ea3e19601d7\""
-	properties.LastModified = &now
-	jsonString, err := json.Marshal(properties)
+	properties.ETag = &etag
+	properties.LastModified = nowAsUnix
+	jsonString, err := json.Marshal(&properties)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -93,17 +92,17 @@ func TestThroughputPropertiesManualE2ESerialization(t *testing.T) {
 		t.Errorf("OfferId mismatch %v", otherProperties.offerId)
 	}
 
-	if otherProperties.ETag != "\"00000000-0000-0000-9b8c-8ea3e19601d7\"" {
+	if *otherProperties.ETag != etag {
 		t.Errorf("Etag mismatch %v", otherProperties.ETag)
 	}
 
-	if otherProperties.LastModified.Time != properties.LastModified.Time {
-		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified.Time)
+	if otherProperties.LastModified != properties.LastModified {
+		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified)
 	}
 
-	mt, err := otherProperties.ManualThroughput()
-	if err != nil {
-		t.Fatal(err)
+	mt, isManual := otherProperties.ManualThroughput()
+	if !isManual {
+		t.Fatal("Expected to have manual throughput available")
 	}
 
 	if mt != 400 {
@@ -111,19 +110,16 @@ func TestThroughputPropertiesManualE2ESerialization(t *testing.T) {
 	}
 }
 
-func TestThroughputPropertiesAutoscaleE2ESerialization(t *testing.T) {
-	nowAsUnix := time.Now().Unix()
+func TestThroughputPropertiesAutoscaleWithIncrementE2ESerialization(t *testing.T) {
+	nowAsUnix := time.Unix(time.Now().Unix(), 0)
 
-	now := UnixTime{
-		Time: time.Unix(nowAsUnix, 0),
-	}
-
-	properties := NewAutoscaleThroughputProperties(400)
+	etag := azcore.ETag("\"00000000-0000-0000-9b8c-8ea3e19601d7\"")
+	properties := NewAutoscaleThroughputPropertiesWithIncrement(400, 500)
 	properties.offerId = "HFln"
 	properties.offerResourceId = "4SRTANCD3Dw="
-	properties.ETag = "\"00000000-0000-0000-9b8c-8ea3e19601d7\""
-	properties.LastModified = &now
-	jsonString, err := json.Marshal(properties)
+	properties.ETag = &etag
+	properties.LastModified = nowAsUnix
+	jsonString, err := json.Marshal(&properties)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -150,17 +146,89 @@ func TestThroughputPropertiesAutoscaleE2ESerialization(t *testing.T) {
 		t.Errorf("OfferId mismatch %v", otherProperties.offerId)
 	}
 
-	if otherProperties.ETag != "\"00000000-0000-0000-9b8c-8ea3e19601d7\"" {
+	if *otherProperties.ETag != etag {
 		t.Errorf("Etag mismatch %v", otherProperties.ETag)
 	}
 
-	if otherProperties.LastModified.Time != properties.LastModified.Time {
-		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified.Time)
+	if otherProperties.LastModified != properties.LastModified {
+		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified)
 	}
 
-	at, err := otherProperties.AutoscaleMaxThroughput()
+	at, hasAutoscale := otherProperties.AutoscaleMaxThroughput()
+	if !hasAutoscale {
+		t.Errorf("Expected to have autoscale")
+	}
+
+	inc, hasAutoscale := otherProperties.AutoscaleIncrement()
+	if !hasAutoscale {
+		t.Errorf("Expected to have autoscale")
+	}
+
+	if at != 400 {
+		t.Errorf("MaxThroughput mismatch %v", at)
+	}
+
+	if inc != 500 {
+		t.Errorf("Increment mismatch %v", inc)
+	}
+
+	if otherProperties.offer.AutoScale.AutoscaleAutoUpgradeProperties == nil {
+		t.Errorf("AutoscaleAutoUpgradeProperties mismatch %v", *otherProperties.offer.AutoScale.AutoscaleAutoUpgradeProperties)
+	}
+}
+
+func TestThroughputPropertiesAutoscaleE2ESerialization(t *testing.T) {
+	nowAsUnix := time.Unix(time.Now().Unix(), 0)
+
+	etag := azcore.ETag("\"00000000-0000-0000-9b8c-8ea3e19601d7\"")
+	properties := NewAutoscaleThroughputProperties(400)
+	properties.offerId = "HFln"
+	properties.offerResourceId = "4SRTANCD3Dw="
+	properties.ETag = &etag
+	properties.LastModified = nowAsUnix
+	jsonString, err := json.Marshal(&properties)
 	if err != nil {
 		t.Fatal(err)
+	}
+
+	otherProperties := &ThroughputProperties{}
+	err = json.Unmarshal(jsonString, otherProperties)
+	if err != nil {
+		t.Fatal(err, string(jsonString))
+	}
+
+	if otherProperties.offerType != "" {
+		t.Errorf("OfferType mismatch %v", otherProperties.offerType)
+	}
+
+	if otherProperties.offerResourceId != "4SRTANCD3Dw=" {
+		t.Errorf("OfferResourceId mismatch %v", otherProperties.offerResourceId)
+	}
+
+	if otherProperties.version != offerVersion2 {
+		t.Errorf("OfferVersion mismatch %v", otherProperties.version)
+	}
+
+	if otherProperties.offerId != "HFln" {
+		t.Errorf("OfferId mismatch %v", otherProperties.offerId)
+	}
+
+	if *otherProperties.ETag != etag {
+		t.Errorf("Etag mismatch %v", otherProperties.ETag)
+	}
+
+	if otherProperties.LastModified != properties.LastModified {
+		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified)
+	}
+
+	at, hasAutoscale := otherProperties.AutoscaleMaxThroughput()
+	if !hasAutoscale {
+		t.Errorf("Expected to have autoscale")
+	}
+
+	_, hasAutoscaleIncrement := otherProperties.AutoscaleIncrement()
+	if hasAutoscaleIncrement {
+		t.Errorf("Expected not to have autoscale increment")
 	}
 
 	if at != 400 {
@@ -168,23 +236,20 @@ func TestThroughputPropertiesAutoscaleE2ESerialization(t *testing.T) {
 	}
 
 	if otherProperties.offer.AutoScale.AutoscaleAutoUpgradeProperties != nil {
-		t.Errorf("AutoscaleAutoUpgradeProperties mismatch %v", otherProperties.offer.AutoScale.AutoscaleAutoUpgradeProperties)
+		t.Errorf("AutoscaleAutoUpgradeProperties mismatch %v", *otherProperties.offer.AutoScale.AutoscaleAutoUpgradeProperties)
 	}
 }
 
 func TestThroughputPropertiesAutoscaleIncrementE2ESerialization(t *testing.T) {
-	nowAsUnix := time.Now().Unix()
+	nowAsUnix := time.Unix(time.Now().Unix(), 0)
 
-	now := UnixTime{
-		Time: time.Unix(nowAsUnix, 0),
-	}
-
+	etag := azcore.ETag("\"00000000-0000-0000-9b8c-8ea3e19601d7\"")
 	properties := NewAutoscaleThroughputPropertiesWithIncrement(400, 10)
 	properties.offerId = "HFln"
 	properties.offerResourceId = "4SRTANCD3Dw="
-	properties.ETag = "\"00000000-0000-0000-9b8c-8ea3e19601d7\""
-	properties.LastModified = &now
-	jsonString, err := json.Marshal(properties)
+	properties.ETag = &etag
+	properties.LastModified = nowAsUnix
+	jsonString, err := json.Marshal(&properties)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -211,17 +276,17 @@ func TestThroughputPropertiesAutoscaleIncrementE2ESerialization(t *testing.T) {
 		t.Errorf("OfferId mismatch %v", otherProperties.offerId)
 	}
 
-	if otherProperties.ETag != "\"00000000-0000-0000-9b8c-8ea3e19601d7\"" {
+	if *otherProperties.ETag != etag {
 		t.Errorf("Etag mismatch %v", otherProperties.ETag)
 	}
 
-	if otherProperties.LastModified.Time != properties.LastModified.Time {
-		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified.Time)
+	if otherProperties.LastModified != properties.LastModified {
+		t.Errorf("Timestamp mismatch %v", otherProperties.LastModified)
 	}
 
-	at, err := otherProperties.AutoscaleMaxThroughput()
-	if err != nil {
-		t.Fatal(err)
+	at, hasAutoscale := otherProperties.AutoscaleMaxThroughput()
+	if !hasAutoscale {
+		t.Errorf("Expected to have autoscale")
 	}
 
 	if at != 400 {